refactor(fingerprint): Shift focus from Metadata to MetaInfo
Now called `UnitHash` and `Metadata`.

The goal is to provide role-specific `UnitHash` accessors on the new `Metadata`.
I chose this method, instead of changing the accessors on
`CompilationFiles`, to have the least impact on the documentation as
what is called `Metadata` can remain the focus of the documentation,
rather than having it apply to different `CompilationFiles` accessors
which don't house the actual logic.