test(msrv): Workaround Edition 2024 stabilization
With 2024 stabilized, there will be a "first supported" version which is
above the versions used in tests.

In particular, for the `cargo install` test, we can't test the scenario
until a couple of versions have passed.
We have to rely on Edition 2024 changing ony resolver v3 which is
already covered.