fix(tree): Make output more deterministic
When you have a dependency show up in multiple dependency tables,
the order they are processed was not deterministic.
Its hard to predict the exact effects this was having (e.g. for writing
tests for this) but changing this made #15366 not reproduce anymore for me.

I'm going to consier this as fixing #15366 mostly because how difficult
it is to debug with the non-determinism.
If someone can find a case where this fails, at least it should now
always fail and it should be easier to determine why.

Fixes #15366