Print all features  when the verbose flag is set
Currently, the dependency printing has a hardcoded limit, and
for some crates it's useful to see a bit more. I understand why
the hard limit is set, but we already have a verbose flag, let's
use it and let the user decide if he wants to see something more.

My case was the crate `reqest`, that listed for me as:

```
➜  cargo add reqwest
    Updating crates.io index
      Adding reqwest v0.12.15 to dependencies
             Features:
             + __tls
             + charset
             + default-tls
             + h2
             + http2
             + macos-system-configuration
             26 deactivated features
```

What are those deactivated features? I had to use the browser
to discover, whereas it's much more confortable for me to just
do a small pipe to grep and `-` to see only the features I care
about: the ones that are disabled, and I might need to add to
my project.

Overall, this is a really small change, but made my life easier.